Split app/admin/users/[userId]/page.tsx into sub-files
